dom/base/test/test_bug405182.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=405182
-->
<head>
  <title>Test for Bug 405182</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=405182">Mozilla Bug 405182</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 405182 **/

function do_test()
{
  var dE = document.documentElement;

  document.addEventListener("DOMNodeRemoved", newScript);

  document.removeChild(dE);

  function newScript()
  {
    var ns = document.createElementNS("http://www.w3.org/1999/xhtml", "script"); 
    var nt = document.createTextNode("42;");
    ns.appendChild(nt);
    dE.appendChild(ns);
    ok(true, "Test is successful if we get here without crashing");
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(do_test);

</script>
</pre>
</body>
</html>
